Fix capitalization of mp3 URLs in anchor tags included in clues.
Clues are automatically converted to upper case before being stored. This inadvertently converts URLs to upper case as well, which would cause them to lead to dead links as URLs are case sensitive. The ".wmv" and ".jpg" extensions are already re-converted to lower case; now, ".mp3" extensions are also re-converted.

- Replace ".MP3" with ".mp3" in clue URLs using String.replace()
- Replace "A." with "a." in clue URLs using String.replace()
- Replace "HTTP" and "J-ARCHIVE" in clue URLs with "http" and "j-archive", respectively, using String.replace()
- Set the target of clue URLs to "_blank" so that they open in a new tab/window