Show the end screen in games where there are fewer than 30 clues.
Certain games have unrevealed clues. This causes the board to have fewer than 30 clues. Initially, the end screen is only triggered when 30 clues have been revealed. To account for cases where there are fewer than 30 clues that can be revealed, keep track of the actual number of clues available on the board.

- Define the variable actualNumberOfClues to keep track of the actual number of non-greyed out clues in the current game